backend/app/utils/config.py
import os
from dotenv import load_dotenv

load_dotenv()


class Config:
    SECRET_KEY = os.getenv('SECRET_KEY', 'change-me')
    MONGO_URI = os.getenv('MONGO_URI', 'mongodb://localhost:27017')
    DB_NAME = os.getenv('DB_NAME', 'pfh')
    JWT_EXP_MINUTES = int(os.getenv('JWT_EXP_MINUTES', '10080'))
